fix(gates): split oldgen sweep-cleanup and the cycle alloc-flag helper for the file cap
